refactor(engine): 外移 status.rs 内联测试至 status_tests.rs
- status.rs 4594→3020 行，纯搬运零行为变更
- #[cfg(test)] #[path] 平铺，参照 commands_tests.rs 先例
- baseline 重生成：status.rs reduced -479；new-file baseline 登记 status_tests.rs
- 顺带冻结 main 存量增长（manager/pi 等），属治理变更
验证：cargo test --lib engine::status 56 passed；cargo check 通过；check:large-files:gate 绿
